requirements.txt
PyNaCl==1.5.0
aiohttp>=3.9.0
ddgs>=9.0.0
discord.py==2.3.2
openai>=1.30.0
python-dotenv>=1.0.0
wikipedia==1.4.0